+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
   stack drift details from script jobs.
 - Add --cloud-sync-layer to allow users to specify a preview layer, e.g.: `stg`, `prod` etc.
   - This is useful when users want to preview changes in a specific terraform workspace.
+- Add `--cloud-sync-layer` and `--cloud-sync-preview` to `script` block, this would allow users to synchronize previews to Terramate Cloud via script jobs.
 
 ### Fixed